avm2: Add ignored test for subclass with 'overriding' field name
The Adobe Animate compiler rejects a subclass that contains
a non-private field with the same name as a field in an ancestor
class (e.g. 'pub var foo:String' in both the subclass and superclass).

Unfortunately, Flash Player accepts this code, and creates a distinct
field for each class (even though they have the same namespace and
name). I suspect that this is caused by the optimizer replacing the
field accesses with internal slot accesses.

I've added an ignored test demonstrating this behavior - getting Ruffle
to reproduce it will be tricky.