render: Implement more Pixelbender opcodes, and fix matrices
Matrices don't actually get their own registers - they are
stored in normal float registers, with each column of
the matrix getting its own register. The exception is float2x2,
which is stored in a single vec4 (this is visible to PixelBender
bytecode, which writes to a single vector, and then uses it
as an M2x2 channel).