avm2: Fix handling of Vector.<*>
We were previously using the Vector.<Object> `Class`
with two different `ClassObjects` (`Vector.<*>` and
`Vector.<Object>`), which resulted in the class initializer
getting skipped for a specialized class.

We also turned `Vector.<*>` into `Vector.<Object>` in
the Vector class call handler, resulting in getQualifiedClassName
printing `Vector.<Object>` instead of `Vector.<*>`. We now
correctly apply the class in the call handler (but still use
a value type of `Object` in `VectorStorage` for `Vector.<*>`)