avm2: Implement DisplayObject.blendShader
We already had almost all of the PixelBender support needed for this.
Since this is run during displayobject rendering, we need to support
MSAA, which requires us to create a `RenderPipeline` per sample count
used.

This fixes Closure's rendering (previously, the game rendered as a solid
white background with a few gray ovals).