avm2: Add support for Stage3D bytearray/compressed textures
'Compressed' textures are Adobe Texture Format (ATF) wrappers
around the obscure JPEG-XR image format. To implement this,
I've forked the Rust 'jpegxr' crate, and added in a method
to convert a JPEG-XR image to a TIFF (which can then be handled
by the standard 'image' crate).

The 'jpegxr' crate wraps the C 'jxrlib' library. In order for this
to work on web, I've modified the crate to include a 'fakelib' impl
with stubs and copy-pasted musl method for various libc methods
(based on szolo's previous work).