core: Use HTTP response encoding if existing
The form loader now uses the encoding specified in the HTTP response
content type field if existing to decode remote text files. chardetng is
now (only) used if the HTTP response doesn't specify any encoding or if
the file is local.

To do this, a text_encoding variable has been added to the
SuccessResponse struct. The fetch implementations parse the header and
set the variable accordingly.