core: Register character 0 with library in Loader
This allows ActionScript code to instantiate the AVM2 class
for the root MovieClip of a loaded movie. This instance will
have its frames run, including timeline-placed children.

We now store `avm2_class` in `MovieClipStatic`, so that processing
a `SymbolClass` tag from a loaded MovieClip will update the class
used by the MovieClip stored in the Library.