core: Key input overhaul
The main goal of this refactor is to unify key input across all
platforms and move all Flash Player related behaviors to core in order
to get rid of differences across platforms and minimize bugs.

Instead of providing KeyCodes in PlayerEvents, frontends now provide
KeyDescriptors, which describe the pressed key on a high level.
Then, Ruffle core does all necessary mappings to KeyCodes
(and key location in the future).