web: Resolve URLs without a protocol in the extension player
Previously, only URLs explicitly including the web protocol (HTTPS or
HTTP) in the beginning could be resolved in the extension player.
This has been changed. The extension player now resolves URLs without a
protocol and adds https:// or http:// if the URL is likely a valid web
URL and file:/// if the URL is a valid file URL.
HTTP is only used if the server only supports HTTP and not HTTPS.