web: Display loading screen while resolving the SWF URL
Previously, when entering a URL in the extension player, it had been
resolved before creating a RufflePlayer instance and displaying a
loading screen. As resolving can include testing for server availability
(to decide whether HTTPS or HTTP is used), this could create a wait time
without the loading screen.
This has been fixed. The URL is now resolved while the loading screen is
displayed. Additionally, the timeouts have been increased (they were low
because of the lack of a loading animation).

A bug that caused the document title to not be set correctly if the SWF
URL was the main page of a domain has also been fixed.