core: Allow playhead to advance past SWF head frame count
The frame count declared in the SWF header is not required to
match the actual number of frames present in the tag stream.
Flash Player will execute frames until it reaches the end of
the tag stream (reporting a `currentFrame` larger both `totalFrames`
and `framesLoaded`), and *then* jump back to the start.

We now inspect the tag stream in `determine_next_frame` - if we've
reached the end (either `TagCode::End` or we've run out of bytes),
we jump back to the beginning. The declared SWF frame count is ignored.

I haven't addressed gotos yet, which are more complicated due to
the 'frame clamping' behavior.